refac : 관리자 셀프소개 패키지 삭제 및 컨트롤러 이동
