+++ src/main/java/atwoz/atwoz/community/presentation/selfintroduction/AdminSelfIntroductionController.java
@@ -1,11 +1,11 @@
-package atwoz.atwoz.admin.presentation.selfintroduction;
+package atwoz.atwoz.community.presentation.selfintroduction;
 
-import atwoz.atwoz.admin.command.application.selfintroduction.AdminSelfIntroductionCommandService;
-import atwoz.atwoz.admin.query.selfintroduction.AdminSelfIntroductionQueryRepository;
-import atwoz.atwoz.admin.query.selfintroduction.AdminSelfIntroductionSearchCondition;
-import atwoz.atwoz.admin.query.selfintroduction.AdminSelfIntroductionView;
 import atwoz.atwoz.common.enums.StatusType;
 import atwoz.atwoz.common.response.BaseResponse;
+import atwoz.atwoz.community.command.application.selfintroduction.SelfIntroductionService;
+import atwoz.atwoz.community.presentation.selfintroduction.dto.AdminSelfIntroductionSearchCondition;
+import atwoz.atwoz.community.query.selfintroduction.SelfIntroductionQueryRepository;
+import atwoz.atwoz.community.query.selfintroduction.view.AdminSelfIntroductionView;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -21,8 +21,8 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class AdminSelfIntroductionController {
 
-    private final AdminSelfIntroductionQueryRepository adminSelfIntroductionQueryRepository;
-    private final AdminSelfIntroductionCommandService adminSelfIntroductionCommandService;
+    private final SelfIntroductionQueryRepository selfIntroductionQueryRepository;
+    private final SelfIntroductionService selfIntroductionService;
 
     @Operation(summary = "셀프 소개 목록 조회 API")
     @GetMapping
@@ -31,20 +31,20 @@ public class AdminSelfIntroductionController {
         @PageableDefault(size = 100) Pageable pageable
     ) {
         return ResponseEntity.ok(BaseResponse.of(StatusType.OK,
-            adminSelfIntroductionQueryRepository.findSelfIntroductions(condition, pageable)));
+            selfIntroductionQueryRepository.findSelfIntroductions(condition, pageable)));
     }
 
     @Operation(summary = "셀프 소개 공개 전환 API")
     @PatchMapping("/{id}/open")
     public ResponseEntity<BaseResponse<Void>> updateSelfIntroductionToOpen(@PathVariable Long id) {
-        adminSelfIntroductionCommandService.convertToOpen(id);
+        selfIntroductionService.changeOpenStatus(id, true);
         return ResponseEntity.ok(BaseResponse.from(StatusType.OK));
     }
 
     @Operation(summary = "셀프 소개 비공개 전환 API")
     @PatchMapping("/{id}/close")
     public ResponseEntity<BaseResponse<Void>> updateSelfIntroductionToClose(@PathVariable Long id) {
-        adminSelfIntroductionCommandService.convertToClose(id);
+        selfIntroductionService.changeOpenStatus(id, false);
         return ResponseEntity.ok(BaseResponse.from(StatusType.OK));
     }
 }
