+++ src/main/java/atwoz/atwoz/community/presentation/selfintroduction/SelfIntroductionController.java
@@ -5,10 +5,10 @@ import atwoz.atwoz.auth.presentation.AuthPrincipal;
 import atwoz.atwoz.common.enums.StatusType;
 import atwoz.atwoz.common.response.BaseResponse;
 import atwoz.atwoz.community.command.application.selfintroduction.SelfIntroductionService;
+import atwoz.atwoz.community.presentation.selfintroduction.dto.SelfIntroductionSearchCondition;
 import atwoz.atwoz.community.presentation.selfintroduction.dto.SelfIntroductionSearchRequest;
 import atwoz.atwoz.community.presentation.selfintroduction.dto.SelfIntroductionWriteRequest;
 import atwoz.atwoz.community.query.selfintroduction.SelfIntroductionQueryRepository;
-import atwoz.atwoz.community.query.selfintroduction.SelfIntroductionSearchCondition;
 import atwoz.atwoz.community.query.selfintroduction.view.SelfIntroductionSummaryView;
 import atwoz.atwoz.community.query.selfintroduction.view.SelfIntroductionView;
 import lombok.RequiredArgsConstructor;
