+++ src/main/java/atwoz/atwoz/community/presentation/selfintroduction/SelfIntroductionMapper.java
@@ -1,7 +1,7 @@
 package atwoz.atwoz.community.presentation.selfintroduction;
 
+import atwoz.atwoz.community.presentation.selfintroduction.dto.SelfIntroductionSearchCondition;
 import atwoz.atwoz.community.presentation.selfintroduction.dto.SelfIntroductionSearchRequest;
-import atwoz.atwoz.community.query.selfintroduction.SelfIntroductionSearchCondition;
 import atwoz.atwoz.member.command.domain.member.Gender;
 import atwoz.atwoz.member.query.member.AgeConverter;
