+++ src/main/java/atwoz/atwoz/community/query/selfintroduction/view/AdminSelfIntroductionView.java
@@ -1,4 +1,4 @@
-package atwoz.atwoz.admin.query.selfintroduction;
+package atwoz.atwoz.community.query.selfintroduction.view;
 
 import com.querydsl.core.annotations.QueryProjection;
